python/pyproject.toml
[tool.poetry]
name = "hedera-agent-kit"
version = "3.1.0"
description = "Use the Hedera Agent Kit to build agents that can interact with the Hedera network in any framework you like"
authors = ["Hashgraph"]
license = "Apache-2.0"
readme = "README.md"
packages = [{ include = "hedera_agent_kit" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
langchain = "==1.2.0"
langchain-openai = "==1.1.6"
langchain-anthropic = "==1.3.0"
langchain-groq = "==1.1.1"
python-dotenv = ">=1,<2"
aiohttp = ">=3.13.1,<4.0.0"
pydantic = "^2.12.3"
hiero-sdk-python = "0.1.9"
web3 = "==7.14.0"

[tool.poetry.group.dev.dependencies]
black = ">=25.9.0,<26.0.0"
pytest = ">=8.4.2,<9.0.0"
mypy = ">=1.18.2,<2.0.0"
ruff = ">=0.14.1,<0.15.0"
pytest-asyncio = "==1.2.0"
pytest-timeout = "==2.4.0"
pytest-rerunfailures = "==16.1"
pytest-mock = "==3.15.1"
pytest-xdist = ">=3.5.0,<4.0.0"
pydeps = "==3.0.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
